requirements.txt
Django==3.1.1
django-import-export==2.5.0
certifi==2020.6.20
dnspython==2.0.0
appdirs==1.4.4
future==0.18.2
install==1.3.4
asgiref==3.2.10
pytz==2020.1
sqlparse==0.3.1